+++ src/components/Carousel/CarouselItem/index.scss
