+++ src/pages/home/sections/OurWork/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { useMediaQuery } from 'usehooks-ts'
 import Button from '../../../../components/Button'
 import Carousel from '../../../../components/Carousel'
-import CarouselItem from '../../../../components/CarouselItem'
+import CarouselItem from '../../../../components/Carousel/CarouselItem'
 import { CarouselSlider_mockData } from '../../../../mockData/carouselSlider_mockData'
 import HeaderSection from '../../../../components/HeaderSection'
 import { app_mockData } from '../../../../mockData/app_mockData'
