+++ src/pages/home/sections/WhyWeAre/WhyWeAreItem/index.scss
